+++ includes/class-github-api.php
@@ -25,11 +25,7 @@ class GitHub_Updater_GitHub_API extends GitHub_Updater {
 	 */
 	public function __construct( $type ) {
 		$this->type  = $type;
-		self::$hours = 4;
-
-		if ( ! empty( $this->type->timeout ) ) {
-			self::$hours = (float) $this->type->timeout;
-		}
+		self::$hours = 12;
 	}
 
 	/**
@@ -44,11 +40,12 @@ class GitHub_Updater_GitHub_API extends GitHub_Updater {
 	 * @return boolean|object
 	 */
 	protected function api( $url ) {
-		$response = wp_remote_get( $this->get_api_url( $url ) );
+		$response      = wp_remote_get( $this->get_api_url( $url ) );
+		$code          = wp_remote_retrieve_response_code( $response );
+		$allowed_codes = array( 200, 404 );
 
-		if ( is_wp_error( $response ) || ( ( '200' || '404' ) != wp_remote_retrieve_response_code( $response ) ) ) {
-			return false;
-		}
+		if ( is_wp_error( $response ) ) { return false; }
+		if ( ! in_array( $code, $allowed_codes, true ) ) { return false; }
 
 		return json_decode( wp_remote_retrieve_body( $response ) );
 	}
@@ -103,13 +100,13 @@ class GitHub_Updater_GitHub_API extends GitHub_Updater {
 	 * @since 1.0.0
 	 */
 	public function get_remote_info( $file ) {
-		$response = get_site_transient( 'ghu-' . md5( $this->type->repo . $file ) );
+		$response = $this->get_transient( $file );
 
 		if ( ! $response ) {
 			$response = $this->api( '/repos/:owner/:repo/contents/' . $file );
 
 			if ( $response ) {
-				set_site_transient( 'ghu-' . md5( $this->type->repo . $file ), $response, ( self::$hours * HOUR_IN_SECONDS ) );
+				$this->set_transient( $file, $response );
 			}
 		}
 
@@ -164,7 +161,7 @@ class GitHub_Updater_GitHub_API extends GitHub_Updater {
 	 * @return string latest tag.
 	 */
 	public function get_remote_tag() {
-		$response = get_site_transient( 'ghu-' . md5( $this->type->repo . 'tags' ) );
+		$response = $this->get_transient( 'tags' );
 
 		if ( ! $response ) {
 			$response = $this->api( '/repos/:owner/:repo/tags' );
@@ -175,7 +172,7 @@ class GitHub_Updater_GitHub_API extends GitHub_Updater {
 			}
 
 			if ( $response ) {
-				set_site_transient( 'ghu-' . md5( $this->type->repo . 'tags' ), $response, ( self::$hours * HOUR_IN_SECONDS ) );
+				$this->set_transient( 'tags', $response );
 			}
 		}
 
@@ -245,28 +242,29 @@ class GitHub_Updater_GitHub_API extends GitHub_Updater {
 	 * Uses a transient to limit calls to the API.
 	 *
 	 * @since 1.9.0
-	 * @return base64 decoded CHANGES.md or false
+	 * @param $changes
+	 *
+	 * @return bool
 	 */
 	public function get_remote_changes( $changes ) {
 		if ( ! class_exists( 'MarkdownExtra_Parser' ) && ! function_exists( 'Markdown' ) ) {
 			require_once 'markdown.php';
 		}
 
-		$response = get_site_transient( 'ghu-' . md5( $this->type->repo . 'changes' ) );
+		$response = $this->get_transient( 'changes' );
 
 		if ( ! $response ) {
 			$response = $this->api( '/repos/:owner/:repo/contents/' . $changes  );
 
 			if ( $response ) {
-				set_site_transient( 'ghu-' . md5( $this->type->repo . 'changes' ), $response, ( self::$hours * HOUR_IN_SECONDS ) );
+				$this->set_transient( 'changes', $response );
 			}
 		}
 
 		if ( ! $response ) { return false; }
 		if ( isset( $response->message ) ) { return false; }
 
-		$changelog = '';
-		$changelog = get_site_transient( 'ghu-' . md5( $this->type->repo . 'changelog' ), $changelog );
+		$changelog = $this->get_transient( 'changelog' );
 
 		if ( ! $changelog ) {
 			if ( function_exists( 'Markdown' ) ) {
@@ -274,7 +272,7 @@ class GitHub_Updater_GitHub_API extends GitHub_Updater {
 			} else {
 				$changelog = '<pre>' . base64_decode( $response->content ) . '</pre>';
 			}
-			set_site_transient( 'ghu-' . md5( $this->type->repo . 'changelog' ), $changelog, ( self::$hours * HOUR_IN_SECONDS ) );
+			$this->set_transient( 'changelog', $changelog );
 		}
 
 		$this->type->sections['changelog'] = $changelog;
@@ -288,14 +286,14 @@ class GitHub_Updater_GitHub_API extends GitHub_Updater {
 	 * @return base64 decoded repository meta data
 	 */
 	public function get_repo_meta() {
-		$response   = get_site_transient( 'ghu-' . md5( $this->type->repo . 'meta' ) );
+		$response   = $this->get_transient( 'meta' );
 		$meta_query = '?q=' . $this->type->repo . '+user:' . $this->type->owner;
 
 		if ( ! $response ) {
 			$response = $this->api( '/search/repositories' . $meta_query );
 
 			if ( $response ) {
-				set_site_transient( 'ghu-' . md5( $this->type->repo . 'meta' ), $response, ( self::$hours * HOUR_IN_SECONDS ) );
+				$this->set_transient( 'meta', $response );
 			}
 		}
 
@@ -312,29 +310,8 @@ class GitHub_Updater_GitHub_API extends GitHub_Updater {
 	 * @since 2.2.0
 	 */
 	private function add_meta_repo_object() {
+		$this->type->rating       = $this->make_rating( $this->type->repo_meta );
 		$this->type->last_updated = $this->type->repo_meta->pushed_at;
-		$this->type->rating = $this->make_rating();
-		$this->type->num_ratings = $this->type->repo_meta->watchers;
+		$this->type->num_ratings  = $this->type->repo_meta->watchers;
 	}
-
-	/**
-	 * Create some sort of rating from 0 to 100 for use in star ratings
-	 * I'm really just making this up, more based upon popularity
-	 *
-	 * @since 2.2.0
-	 * @return integer
-	 */
-	private function make_rating() {
-		$watchers    = $this->type->repo_meta->watchers;
-		$forks       = $this->type->repo_meta->forks;
-		$open_issues = $this->type->repo_meta->open_issues;
-		$score       = $this->type->repo_meta->score; //what is this anyway?
-
-		$rating = round( $watchers + ( $forks * 1.5 ) - $open_issues );
-
-		if ( 100 < $rating ) { return 100; }
-
-		return $rating;
-	}
-
 }