+++ includes/class-github-updater.php
@@ -23,7 +23,7 @@ class GitHub_Updater {
 	 *
 	 * @since 1.0.0
 	 *
-	 * @var array
+	 * @var stdClass
 	 */
 	protected $config;
 
@@ -42,28 +42,26 @@ class GitHub_Updater {
 	 * @var integer
 	 */
 	protected static $hours;
-	 
+
 	/**
-	 * Method to set hooks, called in GitHub_Plugin_Updater::__construct via add_action( 'init'...)
-	 * @todo make filter hook work
+	 * Variable for holding transient ids
 	 *
-	 * @since 2.3.0
-	 *
-	 * @return integer
+	 * @var array
 	 */
-	public static function init_hooks() {
-		return apply_filters( 'github_updater_set_transient_hours', self::$hours );
-	}
+	protected static $transients = array();
 
 	/**
 	 * Add extra header to get_plugins();
 	 *
 	 * @since 1.0.0
+	 * @param $extra_headers
+	 *
+	 * @return array
 	 */
 	public function add_plugin_headers( $extra_headers ) {
 		$ghu_extra_headers = array(
-			'GitHub Plugin URI', 'GitHub Branch', 'GitHub Access Token', 'GitHub Timeout',
-			'Bitbucket Plugin URI', 'Bitbucket Branch', 'Bitbucket Timeout',
+			'GitHub Plugin URI', 'GitHub Branch', 'GitHub Access Token',
+			'Bitbucket Plugin URI', 'Bitbucket Branch',
 			);
 		$extra_headers     = array_merge( (array) $extra_headers, (array) $ghu_extra_headers );
 
@@ -74,13 +72,14 @@ class GitHub_Updater {
 	 * Add extra headers to wp_get_themes()
 	 *
 	 * @since 1.0.0
+	 * @param $extra_headers
 	 *
 	 * @return array
 	 */
 	public function add_theme_headers( $extra_headers ) {
 		$ghu_extra_headers = array(
-			'GitHub Theme URI', 'GitHub Branch', 'GitHub Access Token', 'GitHub Timeout',
-			'Bitbucket Theme URI', 'Bitbucket Branch', 'Bitbucket Timeout',
+			'GitHub Theme URI', 'GitHub Branch', 'GitHub Access Token',
+			'Bitbucket Theme URI', 'Bitbucket Branch',
 			);
 		$extra_headers     = array_merge( (array) $extra_headers, (array) $ghu_extra_headers );
 
@@ -158,9 +157,6 @@ class GitHub_Updater {
 					if ( empty( $headers['GitHub Access Token'] ) ) { break; }
 					$git_repo['access_token'] = $headers['GitHub Access Token'];
 					break;
-				case 'GitHub Timeout':
-					if ( empty( $headers['GitHub Timeout'] ) ) { break; }
-					$git_repo['timeout']      = $headers['GitHub Timeout'];
 			}
 		}
 
@@ -183,10 +179,6 @@ class GitHub_Updater {
 				case 'Bitbucket Branch':
 					if ( empty( $headers['Bitbucket Branch'] ) ) { break; }
 					$git_repo['branch']  = $headers['Bitbucket Branch'];
-					break;
-				case 'Bitbucket Timeout':
-					if ( empty( $headers['Bitbucket Timeout'] ) ) { break; }
-					$git_repo['timeout'] = $headers['Bitbucket Timeout'];
 					break;
 			}
 		}
@@ -235,10 +227,8 @@ class GitHub_Updater {
 			$github_uri        = $theme->get( 'GitHub Theme URI' );
 			$github_branch     = $theme->get( 'GitHub Branch' );
 			$github_token      = $theme->get( 'GitHub Access Token' );
-			$github_timeout    = $theme->get( 'GitHub Timeout');
 			$bitbucket_uri     = $theme->get( 'Bitbucket Theme URI' );
 			$bitbucket_branch  = $theme->get( 'Bitbucket Branch' );
-			$bitbucket_timeout = $theme->get( 'Bitbucket Timeout' );
 
 			if ( empty( $github_uri ) && empty( $bitbucket_uri ) ) {
 				continue;
@@ -258,6 +248,7 @@ class GitHub_Updater {
 						$git_theme['owner']                   = $owner_repo[0];
 						$git_theme['repo']                    = $owner_repo[1];
 						$git_theme['name']                    = $theme->get( 'Name' );
+						$git_theme['theme_uri']               = $theme->get( 'ThemeURI' );
 						$git_theme['author']                  = $theme->get( 'Author' );
 						$git_theme['local_version']           = $theme->get( 'Version' );
 						$git_theme['sections']['description'] = $theme->get( 'Description' );
@@ -269,10 +260,6 @@ class GitHub_Updater {
 					case 'GitHub Access Token':
 						if ( empty( $github_token ) ) { break; }
 						$git_theme['access_token']            = $github_token;
-						break;
-					case 'GitHub Timeout':
-						if ( empty( $github_timeout ) ) { break; }
-						$git_theme['timeout']                 = $github_timeout;
 						break;
 				}
 			}
@@ -293,6 +280,7 @@ class GitHub_Updater {
 						$git_theme['owner']                   = $owner_repo[0];
 						$git_theme['repo']                    = $owner_repo[1];
 						$git_theme['name']                    = $theme->get( 'Name' );
+						$git_theme['theme_uri']               = $theme->get( 'ThemeURI' );
 						$git_theme['author']                  = $theme->get( 'Author' );
 						$git_theme['local_version']           = $theme->get( 'Version' );
 						$git_theme['sections']['description'] = $theme->get( 'Description' );
@@ -300,10 +288,6 @@ class GitHub_Updater {
 					case 'Bitbucket Branch':
 						if ( empty( $bitbucket_branch ) ) { break; }
 						$git_theme['branch']                  = $bitbucket_branch;
-						break;
-					case 'Bitbucket Timeout':
-						if ( empty( $bitbucket_timeout ) ) { break; }
-						$git_theme['timeout']                 = $bitbucket_timeout;
 						break;
 				}
 			}
@@ -321,7 +305,7 @@ class GitHub_Updater {
 	protected function set_defaults( $type ) {
 		$this->$type->remote_version        = '0.0.0';
 		$this->$type->newest_tag            = '0.0.0';
-		$this->$type->download_link         = '';
+		$this->$type->download_link         = null;
 		$this->$type->tags                  = array();
 		$this->$type->rollback              = array();
 		$this->$type->sections['changelog'] = 'No changelog is available via GitHub Updater. Create a file <code>CHANGES.md</code> in your repository.';
@@ -333,6 +317,10 @@ class GitHub_Updater {
 		$this->$type->num_ratings           = 0;
 		$this->$type->transient             = array();
 		$this->$type->repo_meta             = array();
+		$this->$type->watchers              = 0;
+		$this->$type->forks                 = 0;
+		$this->$type->open_issues           = 0;
+		$this->$type->score                 = 0;
 	}
 
 	/**
@@ -410,9 +398,10 @@ class GitHub_Updater {
 	/**
 	 * Add Basic Authentication $args to http_request_args filter hook
 	 *
-	 * @param array $args Existing HTTP Request arguments
+	 * @param      $args
+	 * @param null $type
 	 *
-	 * @return mixed Amended HTTP Request arguments
+	 * @return mixed
 	 */
 	public function maybe_authenticate_http( $args, $type=null ) {
 		$username = null;
@@ -440,6 +429,94 @@ class GitHub_Updater {
 		}
 
 		return $args;
+	}
+
+
+	/**
+	 * Used to set_site_transient and checks/stores transient id in array
+	 *
+	 * @param $id
+	 * @param $response
+	 *
+	 * @return bool
+	 */
+	protected function set_transient( $id, $response ) {
+		$transient = 'ghu-' . md5( $this->type->repo . $id );
+		if ( ! in_array( $transient, self::$transients, true ) ) {
+			self::$transients[] = $transient;
+		}
+		set_site_transient( $transient, $response, ( self::$hours * HOUR_IN_SECONDS ) );
+		return true;
+	}
+
+	/**
+	 * Returns site_transient and checks/stores transient id in array
+	 *
+	 * @param $id
+	 *
+	 * @return mixed
+	 */
+	protected function get_transient( $id ) {
+		$transient = 'ghu-' . md5( $this->type->repo . $id );
+		if ( ! in_array( $transient, self::$transients, true ) ) {
+			self::$transients[] = $transient;
+		}
+		return get_site_transient( $transient );
+	}
+
+
+	/**
+	 * Delete all transients from array of transient ids
+	 *
+	 * @param $type
+	 *
+	 * @return bool
+	 */
+	protected function delete_all_transients( $type ) {
+		$transients = get_site_transient( 'ghu-' . $type );
+		if ( ! $transients ) { return false; }
+
+		foreach ( $transients as $transient ) {
+			delete_site_transient( $transient );
+			$key = array_search( $transient, $transients );
+			unset( $transients[ $key ] );
+		}
+		return true;
+	}
+
+
+	/**
+	 * Create transient of $type transients for force-check
+	 *
+	 * @param $type
+	 */
+	protected function make_force_check_transient( $type ) {
+		delete_site_transient( 'ghu-' . $type );
+		set_site_transient( 'ghu-' . $type , self::$transients, 12 * HOUR_IN_SECONDS );
+		self::$transients = array();
+	}
+
+
+	/**
+	 * Create some sort of rating from 0 to 100 for use in star ratings
+	 * I'm really just making this up, more based upon popularity
+	 *
+	 * @since 2.2.0
+	 * @param $repo_meta
+	 *
+	 * @return float|int
+	 */
+	protected function make_rating( $repo_meta ) {
+		$watchers    = ( empty( $repo_meta->watchers ) ? $this->type->watchers : $repo_meta->watchers );
+		$forks       = ( empty( $repo_meta->forks ) ? $this->type->forks : $repo_meta->forks );
+		$open_issues = ( empty( $repo_meta->open_issues ) ? $this->type->open_issues : $repo_meta->open_issues );
+		$score       = ( empty( $repo_meta->score ) ? $this->type->score : $repo_meta->score ); //what is this anyway?
+
+		$rating = round( $watchers + ( $forks * 1.5 ) - $open_issues + $score );
+
+		if ( 100 < $rating ) { return 100; }
+
+		return $rating;
 	}
 
 }